+++ test/index.js
@@ -3,7 +3,6 @@
 const autoChrome = require('..')
 const devices = require('../device')
 const config = require('./helpers/config')
-const { sleep } = require('./helpers')
 
 const { executablePath, userDataDir } = config
 
@@ -33,15 +32,9 @@ async function main() {
 
    await chrome.newPage('https://m.baidu.com/')
 
-   await sleep(1000)
-
    await chrome.page.click('#index-kw')
 
-   await sleep(500)
-
    await chrome.page.keyboard.type('汽车')
-
-   await sleep(500)
 
    await chrome.keyboard.press("Enter")