+++ test/run.js
@@ -11,11 +11,13 @@ async function main() {
    let chrome = await autoChrome({
       executablePath: "C:/Program Files (x86)/Google/Chrome/Application/chrome.exe",
       userDataDir: "C:/Users/Xiang/AppData/Local/Chromium/User Data/",
-      devtools: true,
+      // devtools: true,
       args: [
          `--user-agent=${userAgent}`,
-         '--start-maximized'
+         '--start-maximized',
+         '--force-device-scale-factor=1.25',
       ],
+      headless: true,
       emulate: {
          viewport,
       }